Add day 7
I've added a slightly hacky wrapper, `IntcodeComputer`, to run through
opcodes. This gives me somewhere to keep track of state, so that I can
pause the computer if it needs input but hasn't been given any.

In the future this will need refactoring, pretty dramatically, to bring
all those opcode methods inside this class or another.