Add day 9
This is a massive refactoring of the intcode computer. The new parameter
mode gave me a reason to finally tidy up the parameter handling, adding
`read` and `write` methods to help abstract it. Those methods made it
straightforward to write dramatically simpler code for each operation,
to the point where it's just done inline in a case statement now.

The tests have changed how they're constructing and calling the code,
but the things being tested are more or less the same.